Update secrets generation section of README
Fixes #48 by adding the Concourse binary method as the first way of
generating secrets. Updates the ssh-keygen method by adding the `-m`
flag to specify what format the key needs to be in.

Removed a reference to this issue:
https://github.com/helm/charts/issues/13019

Issue was fixed and therefore not relevant to the chart anymore.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>